==> 01_cleaning.ipynb <==
# Exploración y limpieza de datos

import pandas as pd

mesur = pd.read_csv('measurements.csv')

mesur.sample(5)

mesur.info()

mesur.shape

mesur.isna().sum() #temp_inside 12 // specials 295 //refill liters + refill gas 375

#como refill liters y refill gas tienen 375 de 388 nulls, prescindimos de ellas
mesur.drop(['refill liters','refill gas'],axis=1,inplace=True)

mesur.specials.value_counts()
#specials tiene más del 75% de datos nulos y la mayoría de loss datos aparencen ya en columnas (rain, ac...)
# también prescindimos de esta columna

mesur.drop('specials',axis=1,inplace=True)

mesur.sample(3)

todos los datos son numéricos menos gas_type, por lo que, estas variables, si son tipo objet, las convertiremos a tipo float o int,dependiendo del caso...

mesur.select_dtypes(include=['object']).columns

mesur.distance = mesur.distance.str.replace(',','.')
mesur.consume = mesur.consume.str.replace(',','.')
mesur.temp_inside = mesur.temp_inside.str.replace(',','.')

mesur.distance = mesur.distance.astype('float')
mesur.consume = mesur.consume.astype('float')
mesur.temp_inside = mesur.temp_inside.astype('float')

mesur.info()

mesur.isna().sum() #sigue habiendo nulos en temp_inside, como son pocos los rellenaremos

mesur.temp_inside.mean(),mesur.temp_inside.mode(),mesur.temp_inside.median()

mesur.temp_inside.fillna(mesur.temp_inside.mean(),inplace = True)
#rellenamos los nulos de temp_inside con la media y comprobamos que no hay práctiamente cambio en la propia media
mesur.temp_inside.mean()

mesur.to_csv('export/mesur_clean.csv', index=False)
