==> 1- Cleaning.ipynb <==
# import libraries
import pandas as pd

# get to know the dataset
# change the dataset separation to ','
df = pd.read_csv('dataset/measurements.csv', decimal = ',')
df.head()

df.info()

**We find a 12 column dataset, with 387 rows. Each row give us information about one different trip, including distance, consume, average speed, temperature inside, temperature outside, special data (about weather), the gas type used, if the air conditioner was on/off, if it was raining or sunny, and  the refilled of gas**

**Columns "refill liters" and "refill gas", and "specials" have a wide range of NAN values**


## Distance

# check info
distance_sorted = sorted(df.distance.unique())

# to float
df.distance = pd.to_numeric(df.distance)
# check nans after convert to float
print('nans = ', df.distance.isna().sum())

## Consume

consume_sorted = sorted(df.consume.unique())

## Average speed in km/h

speed_sorted = sorted(df.speed.unique())

## Temperature inside the car

temp_inside_sorted = sorted(df.temp_inside.unique())
print('Have found', df.temp_inside.isna().sum(), 'empty values in this serie')

print(df.temp_inside.value_counts())

# 3% of our data are null, so we decided to replace it for the average temperature
print(round((df.temp_inside.isna().sum())/(df.shape[0]), 2 ))

mean_temp = round(df.temp_inside.mean(), 2) 
df.temp_inside.fillna(mean_temp, inplace = True)

df.temp_inside.isna().sum()

## Temperature outside

df.temp_outside.unique()

## Specials

df.specials.unique()

print(round((df.specials.isna().sum())/(df.shape[0]), 2 ))

**Specials column doesn't give us relevant information, and the 76% of values are nulls, so we decided to drop it**

## Gas Type

df.gas_type.unique()

df["gas_type"] = df["gas_type"].astype('category')
df["gas_type_coded"] = df["gas_type"].cat.codes

## Air Conditionair

df.AC.unique()

## Rain

df.rain.unique()

## Sun

df.sun.unique()

## Refill liters

df['refill liters'].unique()

print(round((df['refill liters'].isna().sum())/(df.shape[0]), 2 ))

## Refill gas 

df['refill gas'].unique()

print(round((df['refill gas'].isna().sum())/(df.shape[0]), 2 ))

**97% of values in refill liters and refill gas are empty values**

**We decided to drop "refill liters" and "refill gas" columns, as well as "specials"**

df = df.drop(columns = ['specials', 'refill gas', 'refill liters'])

## Resultant dataset after cleaning

df.info()

df.describe()

df.to_csv('dataset/df_clean.csv', index = False)
