==> 1-EDA.ipynb <==
# ANALIS EXPLORATORIO


import pandas as pd

## Cargo los dos data sets y los comparo

df = pd.read_csv("measurements.csv")

df.head()

df2 = pd.read_excel("measurements2.xlsx")

df2.head()

df.shape

df2.shape

df.describe()

df2.describe()

df.info()

df2.info()

### Son la misma informacion en dos formatos distintos, me quedo den el que proviene de formato excel por practicidad, las unidades ya están definidas

## Estudio por columnas, valuores nulos y cada columna para determinar si van ha serme de utilidad

col = df2.columns
col

df2.isna().sum()

relleno las filas que faltan con la temperatura media

df2.temp_inside = df2.temp_inside.fillna(21.929521)

df2.specials.unique()

df2.specials.value_counts()

def snow(x):
    try:
        if 'SNOW'in x.upper():
            return 1
        else: return 0
    except:
        return 0


df2['snow']=df.specials.apply(snow)
df2

df2.drop("specials", axis=1, inplace=True)

estudio de las columnas refill:

df2["refill liters"].unique()

df2["refill gas"].unique()

df2[df2["refill liters"]==37.6]

df2[df2["refill liters"]==38.]

df2["refill liters"] = df2["refill liters"].fillna(0)
df2["refill gas"] = df2["refill gas"].fillna("No")


df2.head()

df2.gas_type.unique()

df2.info()

## Despues de visualizar los datos determino crear una columna para calcular el tiempo del recorrido para hacer el data framen más comprensibe para humanos 

df2['time'] =  df2.distance/df2.speed

df2.head()

# Salvo los datos para countinuar en otro archivo

df2.to_csv("datos_limpios.csv", index=False)
