==> 1. Cleaning.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import LinearRegression
from cleaning import * 


df_gas = pd.read_csv("./data/measurements.csv", sep=",", decimal=",")


df_gas

df_gas.columns = df_gas.columns.str.strip().str.lower()
df_gas.head(5)

df_gas.isnull().sum()


df_gas.shape

 df_gas = df_gas.fillna(0)

As we can see, the columns that follow 'specials' contain all the information that 'specials' contains, so that, in order to avoid duplicate information, we proceed to eliminate this 'specials' column.

df_gas.drop('specials',
  axis='columns', inplace=True)

df_gas.shape

Now with fill the NaN values of temp_inside with the mean of the column data

df_gas.dtypes

df_gas['temp_inside'].fillna(value = df_gas['temp_inside'].mean().round(3), inplace=True)

df_gas

Now let's see what's up with the NaN values of the two last columns. The first step is to rename them so they don't have any blank spaces.

df_gas = df_gas.rename(columns = {"refill liters":"refill_liters", "refill gas":"refill_gas"})

df_gas

del df_gas['refill_liters']
del df_gas['refill_gas']

df_gas

We add two new columns formed by the cummulative number of km of distance and the cummulative number of liters consumed

df_gas['km_total'] = df_gas['distance'].cumsum()
df_gas['consume_liter_total'] = df_gas['consume']/100*df_gas['distance']

df_gas.head()

print('DISTANCE:')
gas_type_sum(df_gas, 'E10','distance')
gas_type_std(df_gas, 'E10','distance')
gas_type_mean(df_gas, 'E10','distance')
gas_type_sum(df_gas, 'SP98','distance')
gas_type_std(df_gas, 'SP98','distance')
gas_type_mean(df_gas, 'SP98','distance')
print('\n')
print(100*'*')
print('\n')
print('SPEED:')
gas_type_std(df_gas, 'E10','speed')
gas_type_mean(df_gas, 'E10','speed')
gas_type_std(df_gas, 'SP98','speed')
gas_type_mean(df_gas, 'SP98','speed')
print('\n')
print(100*'*')
print('\n')
print('CONSUME:')
gas_type_mean(df_gas, 'E10','consume')
gas_type_mean(df_gas, 'SP98','consume')

Now, in order to change all categorical variables to numerical, we'll change the gas_type column to boolean:

df_gas['gas_type'] = df_gas['gas_type'].map({'SP98': 1, 'E10': 0}) 
#We've changed string to integer

sp98_price = 1.46
E10_price = 1.38
df_gas['price_per_travel'] = df_gas.consume * df_gas.gas_type.apply(lambda x: E10_price if x == 0 else sp98_price)

df_gas.head()

sorted_columns = ['distance', 'consume', 'speed','temp_inside', 'temp_outside', 'gas_type', 'ac', 'rain', 'sun', 'km_total','consume_liter_total', 'price_per_travel']
df_fas = df_gas.reindex(sorted_columns, axis=1)

df_gas.head()

df_gas.to_csv('./data/clean_measure.csv', index=False)
