==> 1_cleaning.ipynb <==
# Exploring data

import pandas as pd

data1 = pd.read_csv("input/measurements.csv")
data1.head()

data2 = pd.read_csv("input/measurements2.csv", sep = ";")
data2.head()

data1.info(), data2.info()

Data 1 and data 2 are the same, however data 2 types are more convenient since they are converted to float and it is easier to operate with them.

df = pd.read_csv("input/measurements2.csv", sep = ";")

## Null Values

null_cols = df.isnull().sum()
null_cols[null_cols > 0]

- Refill liters column seems useless but we might add them when enriching the data so, let's set it aside at the moment.
- Refill gas column is duplicated since gas_type column contain more detailed information. We can delete it.

df.drop(columns=['refill gas'], inplace=True)

## Standardize data

### Specials

df.specials.value_counts()

**Since there is a speficic column for AC, sun or rain we don't really need to have that data duplicated, however, we will leave it there at the moment**

df.specials = df.specials.replace(["half rain half sun", "sun" ], "sun")
df.specials = df.specials.replace(["rain", "snow" ], "rain")
df.specials = df.specials.replace(["ac", "AC" ], "AC")
df.specials = df.specials.replace(["AC rain", "AC snow","ac rain" ], "AC and rain")
df.specials = df.specials.replace(["sun ac", "AC sun","AC Sun" ], "AC and sun")

df.specials.value_counts()

df.head()

# Exterior Temperature
I consider important to know if it is cold or not. The criteria is that it is cold if is the temperature is cooler outside than inside.

ext = []
'''for o,i in df.temp_outside, df.temp_inside:
    if o>i:
        ext.append('HOT')
    else:
        ext.append('COLD')'''
df["external_temp"] = df.temp_outside - df.temp_inside

for value in df["external_temp"]:
    if value <= 0:
        df["external_temp"] = df["external_temp"].replace(value, "COLD")
    else:
        df["external_temp"] = df["external_temp"].replace(value, "HOT")

## This dataset is ready for its first plots

df

df.to_csv("input/clean_measures.csv")
