==> 1_Data_Exploration_&_Basic_Cleaning.ipynb <==
import pandas as pd

### Cargamos los datos.

data = pd.read_csv('data/measurements.csv')

data.head()

data2 = pd.read_excel('data/measurements2.xlsx')

data2.head()

Nos damos cuenta de que los dos datasets son muy similares, así que lo comprobamos.

data2 == data

Los DataFrames no son exactamente iguales. Parece que lo que ocurre es que contienen la misma información, pero en diferente tipo de datos (dígitos en string vs. float, por ejemplo).

Profundizaremos un poco más en las características de los datos y, si comprobamos que nuestras sospechas se cumplen, nos quedaremos con el dataset que más nos convenga (el que requiera de una limpieza menor para empezar a trabajar).

### Exploración y limpieza básica inicial.

Iremos chequeando lo mismo para los dos datasets para poder compararlos.

data.shape

data2.shape

data.info()

data2.info()

Observamos como los dos datasets tienen el mismo shape y las mismas columnas.

La diferencia entre ambos es el tipo de datos que contienen cada columna. Mientras en data2 las columnas numéricas son de tipo float; en data esta columnas son tipo object (contienen strings). Todo indica a que este dataset es una variación del primero como resultado de diferentes procesos de limpieza básica. Evidentemente nos interesa tener ya los datos en el tipo adecuado, por lo que trabajaremos sobre **data2**.

data2.info()

Encontramos dos columnas con muy pocos valores no nulos (13 sobre 388; 3,35% de los datos).

Probablemente lo más apropiado sea eliminar estas columnas ('refill liters' y 'refill gas') ya que no tenemos manera de manejar los datos nulos).
También encontramos muchos valores nulos en la columna 'specials', pero la proporción sobre el total de registros no es tan elevada, por lo que deberemos estudiar mejor la situación respecto a esta columna.

data2['refill liters'].value_counts()

data2['refill gas'].value_counts()

# Eliminamos estas 2 columnas ya que no podemos identificar el valor que aportan al dataset.

data2.drop(columns = ['refill liters', 'refill gas'], inplace=True)

**Analizamos la situación de 'specials'.**

data2.specials.value_counts()

data2.columns

data2.AC.value_counts()

Nos damos cuenta de que los valores no nulos de esta columna se relacionan con otras columnas del DataFrame. ('AC', 'rain' y 'sun').

Lo que se ha hecho (antes de entregarnos los datos) ha sido generar variables ficticias (dummies) en función del contenido de 'specials': el valor de estas columnas es 0 o 1 en función de si aparece 'AC', 'rain' o 'sun' en la columna 'specials'.

Por tanto, podemos eliminar la columna 'specials' ya que la información que nos aporta está recogida en los datos de las tres columnas anteriormente mencionadas.

data2.drop('specials', axis=1, inplace=True)

data2.head()

data2.info()

Llegados a este punto solo nos quedan 12 valores nulos en la columna 'temp_inside'. Esto probablemente se deba al desconocimiento del dato para los registros con valores nulos. Por el momento no eliminaremos estos registros y simplemente sustituiremos los 'NaN' por 'unknown'.

data2.temp_inside.fillna('unknown', inplace=True)

data2.info()

Ya tenemos los datos preparados para comenzar a trabajar con ellos y tratar de extraer conclusiones relevantes.

Expotamos los datos para trabajar con ellos en otro notebook.

data2.to_csv('data/clean_data.csv', index=False)
