==> 3- Webscraping prices.ipynb <==
It has been gived to us the **general** price for E10 and SP98
- E10 is sold for 1,38€/liter
- SP98 is sold for 1,46€

### However, as we want to do a real comparison and get to know which one do we should use, could be interesting know the *current price* of each one in Spain.

# import libraries
import requests
from bs4 import BeautifulSoup
import pandas as pd
from urllib.request import urlopen
import re
import seaborn as sns
import matplotlib.pyplot as plt

def gasoil(url):
    response = requests.get(url)
    soup = BeautifulSoup(response.content,"html.parser")
    scrap = soup.find_all("table", {'class':'panel tabla_precio_hoy' })
    price = [s.getText().split() for s in scrap]
    # getting the mean, min and max prices for this gasoil
    medio = price[0][3]
    minimo = price[0][4]
    maximo = price[0][5]

    return list((re.findall('\d+\.\d+', medio), re.findall('\d+\.\d+', minimo), re.findall('\d+\.\d+', maximo)))


## Prices for SP98

SP98 = gasoil(url = 'https://www.clickgasoil.com/c/precio-gasolina-98')
print('Mean price today is', SP98[0])
print('Min price today is', SP98[1])
print('Max price today is', SP98[2])

## Prices for E10

E10 = gasoil(url = 'https://www.clickgasoil.com/c/precio-gasolina-95')
print('Mean price today is', E10[0])
print('Min price today is', E10[1])
print('Max price today is', E10[2])

# Comparative of consume and current prices

df = pd.read_csv('dataset/df_clean.csv')
means = df.groupby('gas_type').mean()

mean_E10 = round((means.consume[0])*float(E10[0][0]),2)
min_E10 = round((means.consume[0])*float(E10[1][0]),2)
max_E10 = round((means.consume[0])*float(E10[2][0]),2)
mean_SP98 = round((means.consume[0])*float(SP98[0][0]),2)
min_SP98 = round((means.consume[0])*float(SP98[1][0]),2)
max_SP98 = round((means.consume[0])*float(SP98[2][0]),2)

print('Mean consume with E10  price today is', mean_E10)
print('Min  consume with E10  price today is', min_E10)
print('Max  consume with E10  price today is', max_E10)
print('Mean consume with SP98 price today is', mean_SP98)
print('Min  consume with SP98 price today is', min_SP98)
print('Max  consume with SP98 price today is', max_SP98)

details = {
    'gas type': ['E10', 'SP98'],
    'consume_price_min': [min_E10, min_SP98],
    'consume_price_mean': [mean_E10, mean_SP98],
    'consume_price_max': [max_E10, max_SP98],
    'price_min':[E10[1][0], SP98[1][0]],
    'price_mean':[E10[0][0], SP98[0][0]],
    'price_max':[E10[2][0], SP98[2][0]]
}  

gas = pd.DataFrame(details)
gas

dif_min = round((min_SP98/min_E10)-1, 2)
dif_mean = round((mean_SP98/mean_E10)-1, 2)
dif_max = round((max_SP98/max_E10)-1, 2)
print(dif_min, dif_mean, dif_max)

## In all cases price/consume of E10 is lower.
## SP98 increases price in, at least, a 10% respect to the E10
## We recomend to use E10

# Data enrichment

price_min = dict(zip(gas['gas type'], gas['price_min']))
price_mean = dict(zip(gas['gas type'], gas['price_mean']))
price_max = dict(zip(gas['gas type'], gas['price_max']))

df['price_min'] = df.gas_type.map(price_min)
df['price_mean'] = df.gas_type.map(price_mean)
df['price_max'] = df.gas_type.map(price_max)

df.price_min = pd.to_numeric(df.price_min)
df.price_mean = pd.to_numeric(df.price_mean)
df.price_max = pd.to_numeric(df.price_max)

df['consume_price_min'] = round(df.price_min * df.consume, 2)
df['consume_price_mean'] = round(df.price_mean * df.consume, 2)
df['consume_price_max'] = round(df.price_max * df.consume, 2)

df

df.to_csv('dataset/df_rich.csv', index = False)
