==> 3-Prediccion.ipynb <==
# Predicción con modelos de ml

import pandas as pd
from sklearn.model_selection import train_test_split as tts
from sklearn.linear_model import LinearRegression as LinReg

from sklearn import metrics
import numpy as np

## Preparamos los datos

df = pd.read_csv("datos_limpios.csv")

dicc_gas = {'E10':1,'SP98':0}

df['gas_num'] = df.gas_type.map(dicc_gas)

### Dividimos el data set en test y train y variable objetivo __CONSUMO__


X = df.drop(["consume","gas_type","refill gas"], axis=1)

X.head()

y = df.consume

y.head()

X_train, X_test, y_train, y_test = tts(X,y, test_size=0.2)

### Regresión lineal

linreg=LinReg()

linreg

linreg.fit(X_train, y_train)

linreg.coef_

linreg.intercept_

y_pred = linreg.predict(X_test)

print('MAE - Error Medio Absoluto', metrics.mean_absolute_error(y_test, y_pred))
print('MSE - Error Cuadratico Medio', metrics.mean_squared_error(y_test, y_pred))
print('RMSE - Raiz Error Cuadratico Medio', np.sqrt(metrics.mean_squared_error(y_test, y_pred)))
print('R2 - Coeficiente de Determinacion', metrics.r2_score(y_test, y_pred))

### Intentamos otros modelos

from sklearn.linear_model import Ridge, Lasso
from sklearn.linear_model import SGDRegressor
from sklearn.neighbors import KNeighborsRegressor
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.svm import SVR

models={
    'linreg': LinReg(),
    'ridge': Ridge(),
    'lasso': Lasso(),
    'sgd': SGDRegressor(),
    'knn': KNeighborsRegressor(),
    'grad': GradientBoostingRegressor(),
}

for name, model in models.items():
    print("ENTRENANDO: ", name)
    model.fit(X_train, y_train)

for name, model in models.items():
    y_pred = model.predict(X_test)
    print(f"------{name}------")
    print('MAE - ', metrics.mean_absolute_error(y_test, y_pred))
    print('MSE - ', metrics.mean_squared_error(y_test, y_pred))
    print('RMSE - ', np.sqrt(metrics.mean_squared_error(y_test, y_pred)))
    print('R2 - ', metrics.r2_score(y_test, y_pred))

### Como conclusión no parece viable predecir con precisión el consumo de manera general. Veamos con los datos segun longitud de trayecto.

cd = pd.read_csv("cd_limpios.csv")
md = pd.read_csv("md_limpios.csv")
ld = pd.read_csv("ld_limpios.csv")

cd['gas_num'] = cd.gas_type.map(dicc_gas)
md['gas_num'] = md.gas_type.map(dicc_gas)
ld['gas_num'] = ld.gas_type.map(dicc_gas)

# cd
### Dividimos el data set en test y train y variable objetivo __CONSUMO__


X = cd.drop(["consume","gas_type","refill gas"], axis=1)

X.head()

y = cd.consume

y.head()

X_train, X_test, y_train, y_test = tts(X,y, test_size=0.2)

models={
    'linreg': LinReg(),
    'ridge': Ridge(),
    'lasso': Lasso(),
    'sgd': SGDRegressor(),
    'knn': KNeighborsRegressor(),
    'grad': GradientBoostingRegressor(),
}

for name, model in models.items():
    print("ENTRENANDO: ", name)
    model.fit(X_train, y_train)

for name, model in models.items():
    y_pred = model.predict(X_test)
    print(f"------{name}------")
    print('MAE - ', metrics.mean_absolute_error(y_test, y_pred))
    print('MSE - ', metrics.mean_squared_error(y_test, y_pred))
    print('RMSE - ', np.sqrt(metrics.mean_squared_error(y_test, y_pred)))
    print('R2 - ', metrics.r2_score(y_test, y_pred))
