==> 3. Model choosing.ipynb <==
import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'

import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import LinearRegression
from cleaning import * 
from modeling import *

from sklearn.metrics import mean_squared_error, mean_absolute_error 
from sklearn.metrics import accuracy_score, precision_recall_fscore_support, confusion_matrix 
from sklearn.model_selection import train_test_split 
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import r2_score
from sklearn.svm import SVR
from sklearn import preprocessing
from sklearn.preprocessing import quantile_transform
from sklearn.preprocessing import StandardScaler, MinMaxScaler



import scipy.stats as st
from scipy.stats import pearsonr


from sklearn.linear_model import LinearRegression, Ridge
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.model_selection import cross_validate
import xgboost as xgb


df_model = pd.read_csv("./data/clean_measure.csv")
df_model.head()

df_model = df_model.drop(['temp_inside','km_total', 'consume_liter_total', 'price_per_travel'], axis=1)

df_model

Based on what has been seen in the previous notebooks, we are going to keep the variables: temp_outside, speed, distance and consume:

E10 = df_model[df_model.gas_type == 0]
E10.drop(['gas_type'], axis=1, inplace=True)

SP98 = df_model[df_model.gas_type == 1]
SP98.drop(['gas_type'], axis=1, inplace=True)

cv = 8

X_train_E10, X_test_E10, y_train_E10, y_test_E10 = tr_test_split(E10, 160)


X_train_SP98, X_test_SP98, y_train_SP98, y_test_SP98 = tr_test_split(SP98, 228)

training(X_train_E10, y_train_E10, cv).sort_values('RMSE')

training(X_train_SP98,y_train_SP98, cv).sort_values('RMSE')

As we see the best model of regression is Gradient Boost Regression.

E10.to_csv('./output/E10.csv', index=False)
SP98.to_csv('./output/SP98.csv', index=False)
