==> 3_normalize_data.ipynb <==
import pandas as pd
from sklearn.preprocessing import StandardScaler


df = pd.read_csv("input/clean_measures.csv", index_col = 0)
df.head()

Specials is duplicated in AC, rain and sun, Refill liters only has 13/388 values. 

df.drop(columns=['specials', 'refill liters'], inplace=True)

df.head()

standard = df[['distance','speed','temp_inside','temp_outside']]
dummies = df[['gas_type','external_temp']]

# Standardization

Machine learning algorithms don't perform well when the input numerical attributes have very different scales.

Standarization takes the max-min of each column and put it on the same range.

#Initialize
scaler = StandardScaler()
scaled = scaler.fit_transform(standard)
scaled = pd.DataFrame(scaled,columns=standard.columns)
scaled.head()

scaled.info()

# Get Dummies + Standard

dum = pd.get_dummies(dummies)
dum.head()

scaled_dum = scaler.fit_transform(dum)
scaled_dum = pd.DataFrame(scaled_dum,columns=dum.columns)
scaled_dum.head()

# Final dataset

result = pd.concat([scaled, scaled_dum,df.consume], axis=1)
result.head()

result.to_csv("input/normal_measures.csv")
