==> 4. Gradient Boost Regression.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import RandomizedSearchCV, train_test_split, GridSearchCV
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.preprocessing import quantile_transform
from modeling import *
from sklearn.model_selection import cross_val_score
from sklearn.metrics import mean_squared_log_error,mean_squared_error, r2_score,mean_absolute_error 
import pickle






df_E10 = pd.read_csv('./output/E10.csv')
df_SP98 = pd.read_csv('./output/SP98.csv')


## E10:

X_train_E10, X_test_E10, y_train_E10, y_test_E10 = tr_test_split(df_E10)

model = GradientBoostingRegressor()

param_grid = {
    'max_depth': [20, 30, 40, 50, 60],
    'min_samples_split': [4,5,6,7,8,9],
    'n_estimators': [600, 700, 750, 760, 800, 1000],
    'max_features': ['sqrt']
}


grid_search_E10 = GridSearchCV(estimator = model, param_grid = param_grid, cv = 8, n_jobs = -1, verbose = 2)

model_grid = grid_search_E10.fit(X_train_E10, y_train_E10)

best_model_E10 = model_grid.best_params_
best_model_E10

y_pred_E10 = model_grid.predict(X_test_E10)
print("MSE", round(mean_squared_error(y_test_E10,y_pred_E10)),2)


filename = './models/E10_model.pkl'
with open(filename, 'wb') as file:
    pickle.dump(best_model_E10, file)

## SP98:

X_train_SP98, X_test_SP98, y_train_SP98, y_test_SP98 = tr_test_split(df_SP98)

model = GradientBoostingRegressor()

param_grid = {
    'max_depth': [20, 30, 40, 50, 60],
    'min_samples_split': [4,5,6,7,8,9],
    'n_estimators': [600, 700, 750, 760, 800, 1000],
    'max_features': ['sqrt']
}


grid_search_SP98 = GridSearchCV(estimator = model, param_grid = param_grid, cv = 8, n_jobs = -1, verbose = 2)

model_grid = grid_search_SP98.fit(X_train_SP98, y_train_SP98)

best_model_SP98 = model_grid.best_params_
best_model_SP98

y_pred_SP98 = model_grid.predict(X_test_SP98)
print("MSE", round(mean_squared_error(y_test_SP98,y_pred_SP98)),2)



filename = './models/SP98_model.pkl'
with open(filename, 'wb') as file:
    pickle.dump(best_model_SP98, file)

#### We save the models with the best params in order to use them for the prediction

#filename = './models/E10_model.pkl'
#with open(filename, 'rb') as file:  
 #   E10_model = pickle.load(file)

### E10 prediction:

model_predict = GradientBoostingRegressor(max_depth=20,
 max_features='sqrt',
 min_samples_split=9,
 n_estimators=760) 
model_predict = model_predict.fit(X_train_E10, y_train_E10)
y_pred = model_predict.predict(X_test_E10)

y_test_E10['consume'] = y_pred
#sub = predict[['id','price']]
#sub.to_csv('subm_.csv', index=False, header=True)

y_test_E10['consume']

y_test_E10['consume'].mean()

pr_E10 = y_test_E10[['consume']]
pr_E10.to_csv('./output/consume_prediction_E10.csv', index=False, header=True)

### SP98 prediction:

model_predict = GradientBoostingRegressor(max_depth=20,
 max_features='sqrt',
 min_samples_split=9,
 n_estimators=760) 
model_predict = model_predict.fit(X_train_SP98, y_train_SP98)
y_pred = model_predict.predict(X_test_SP98)

y_test_SP98['consume'] = y_pred
#sub = predict[['id','price']]
#sub.to_csv('subm_.csv', index=False, header=True)

y_test_SP98['consume']

y_test_SP98['consume'].mean()

pr_SP98 = y_test_SP98[['consume']]
pr_SP98.to_csv('./output/consume_prediction_E10.csv', index=False, header=True)

As we can see, the mean of the consume predicted is a little bit smaller in the E10 gas type. If we refresh our memory, we can remember that this gas type was also 0,34 cents per liter cheaper
