==> 5_API_example.ipynb <==
import requests,json
from pymongo import MongoClient
from flask import Flask, render_template, request
from bson.objectid import ObjectId


res = requests.get("http://localhost:5000/users")
a =res.json()
print(list(a.values())[0])
